+++ spinta/manifests/tabular/components.py
@@ -297,6 +297,9 @@ class CommentRow(TypedDict, total=False):
     # TODO: should be datetime
     created: str
     comment: str
+    prepare: str
+    level: str
+    uri: str
 
 
 class CommentData(TypedDict, total=False):